+++ Src/World/HeightMap.cpp
@@ -38,13 +38,13 @@ namespace t3d
 	}
 
 
-	Uint8 HeightMap::get(Uint index)
+	Float HeightMap::get(Uint index)
 	{
 		return mHeightData.at(index);
 	}
 
 
-	Uint8 HeightMap::get(Uint indexX, Uint indexY)
+	Float HeightMap::get(Uint indexX, Uint indexY)
 	{
 		return mHeightData.at(indexY*mSize + indexX);
 	}
